+++ src/shared/kmc-shared/kmc-views/kmc-views.module.ts
@@ -3,6 +3,7 @@ import { KmcMainViewsService } from './kmc-main-views.service';
 import {
     AdminRolesMainViewService,
     AdminUsersMainViewService,
+    AdminMultiAccountMainViewService,
     ContentBulkUploadsMainViewService,
     ContentCategoriesMainViewService,
     ContentDropFoldersMainViewService,
@@ -42,6 +43,7 @@ import {
     LiveDashboardAppViewService,
 } from './component-views';
 import { QuizAppViewService } from 'app-shared/kmc-shared/kmc-views/component-views';
+import { HotspotsAppViewService } from 'app-shared/kmc-shared/kmc-views/component-views/hotspots-app-view.service';
 
 @NgModule({
     imports: <any[]>[],
@@ -71,6 +73,7 @@ export class KmcViewsModule {
                 LiveAnalyticsMainViewService,
                 AdminUsersMainViewService,
                 AdminRolesMainViewService,
+                AdminMultiAccountMainViewService,
                 SettingsAccountSettingsMainViewService,
                 SettingsIntegrationSettingsMainViewService,
                 SettingsAccessControlMainViewService,
@@ -87,6 +90,7 @@ export class KmcViewsModule {
                 ClipAndTrimAppViewService,
                 AdvertisementsAppViewService,
                 QuizAppViewService,
+                HotspotsAppViewService,
                 RestorePasswordViewService,
                 ReachAppViewService,
                 ServicesDashboardMainViewService,